Moved AboutDialog into Dialogs subdir.
