+++ src/GUI/Dialogs/AboutDialog.hpp
